Update style.css
added media query which checks if the screen width is 768 pixels or less.When the screen is narrow (e.g., on mobile devices), it sets the flex-direction of the #app-container to column, which stacks the child elements vertically. It also sets the width of both sidebars to 100% so that they take up the full width of the screen.